R data analysis project
This project was created to showcase my data analysis skills using R.  In this case study, I worked for a fictional company, Cyclistic, along with some key team members. The business question that I was tasked with was to understand the difference in the two types of customers/riders of this company (subscribed members vs. pay-as-you-ride riders).  The main objective of this project was to show that I can follow the data analysis process: Ask, Prepare, Process, Analyze, Share, and Act.